finalização tarefa

+++ GestranAPI/Migrations/ContextGestranModelSnapshot.cs
@@ -64,8 +64,8 @@ namespace GestranAPI.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<int>("CNPJ")
-                        .HasColumnType("int");
+                    b.Property<long>("CNPJ")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Email")
                         .HasColumnType("nvarchar(max)");
